+++ Connectivity/Classes/Connectivity.swift
@@ -51,12 +51,12 @@ public class Connectivity: NSObject {
     public var connectivityCheckLatency: Double = 0.5
     
     /// URLs to contact in order to check connectivity
-    public var connectivityURLs: [URL] = Connectivity
-        .defaultConnectivityURLs(shouldUseHTTPS: Connectivity.isHTTPSOnly) {
+    public var connectivityURLRequests: [URLRequest] = Connectivity
+        .defaultConnectivityURLRequests(shouldUseHTTPS: Connectivity.isHTTPSOnly) {
             didSet {
                 if Connectivity.isHTTPSOnly { // if HTTPS only set only allow HTTPS URLs
-                    connectivityURLs = connectivityURLs.filter { url in
-                        return url.absoluteString.lowercased().starts(with: "https")
+                    connectivityURLRequests = connectivityURLRequests.filter { request in
+                        return request.url?.absoluteString.lowercased().starts(with: "https") ?? false
                     }
                 }
             }
@@ -313,7 +313,7 @@ private extension Connectivity {
     /// Returns a URL request for an Authorization header if `authorizationHeader` or `bearerToken` property is set,
     /// `authorizationHeader` is prefered over `bearerToken`
     /// otherwise `nil` is returned.
-    func authorizedURLRequest(with url: URL) -> URLRequest? {
+    func authorizedURLRequest(with urlRequest: URLRequest) -> URLRequest? {
         var authHeader: String? = authorizationHeader
         if let bearerToken {
             authHeader = authHeader ?? "Bearer \(bearerToken)"
@@ -321,7 +321,7 @@ private extension Connectivity {
         guard let authHeader else {
             return nil
         }
-        var request = URLRequest(url: url)
+        var request = urlRequest
         request.setValue(authHeader, forHTTPHeaderField: "Authorization")
         return request
     }
@@ -351,16 +351,16 @@ private extension Connectivity {
         let dispatchGroup = DispatchGroup()
         var tasks: [URLSessionDataTask] = []
         var successfulChecks: UInt = 0, failedChecks: UInt = 0
-        let totalChecks = UInt(connectivityURLs.count)
+        let totalChecks = UInt(connectivityURLRequests.count)
         
         // Ensure that we are using the latest session configuration.
         urlSession = defaultURLSession()
         
         // Connectivity check callback
-        let completionHandlerForUrl: (URL) -> ((Data?, URLResponse?, Error?) -> Void) = { url in
+        let completionHandlerForURLRequest: (URLRequest) -> ((Data?, URLResponse?, Error?) -> Void) = { urlRequest in
             return { [weak self] data, response, _ in
                 let connectivityCheckSuccess = self?.connectivityCheckSucceeded(
-                    for: url,
+                    for: urlRequest,
                     response: response,
                     data: data
                 ) ?? false
@@ -376,11 +376,9 @@ private extension Connectivity {
         }
         
         // Check each of the specified URLs in turn
-        tasks = connectivityURLs.map {
-            guard let urlRequest = authorizedURLRequest(with: $0) else {
-                return urlSession.dataTask(with: $0, completionHandler: completionHandlerForUrl($0))
-            }
-            return urlSession.dataTask(with: urlRequest, completionHandler: completionHandlerForUrl($0))
+        tasks = connectivityURLRequests.map { request in
+            let urlRequest = authorizedURLRequest(with: request) ?? request
+            return urlSession.dataTask(with: urlRequest, completionHandler: completionHandlerForURLRequest(request))
         }
         
         tasks.forEach { task in
@@ -409,7 +407,7 @@ private extension Connectivity {
 #if canImport(UIKit)
         checkWhenApplicationDidBecomeActive = configuration.checkWhenApplicationDidBecomeActive
 #endif
-        connectivityURLs = configuration.connectivityURLs
+        connectivityURLRequests = configuration.connectivityURLRequests
         externalQueue = configuration.callbackQueue
         framework = configuration.framework
         internalQueue = configuration.connectivityQueue
@@ -425,13 +423,13 @@ private extension Connectivity {
     }
     
     /// Determines whether or not the connectivity check was successful.
-    private func connectivityCheckSucceeded(for url: URL, response: URLResponse?, data: Data?) -> Bool {
+    private func connectivityCheckSucceeded(for urlRequest: URLRequest, response: URLResponse?, data: Data?) -> Bool {
         let validator = responseValidatorFactory.manufacture()
-        return validator.isResponseValid(url: url, response: response, data: data)
+        return validator.isResponseValid(urlRequest: urlRequest, response: response, data: data)
     }
     
     /// Set of connectivity URLs used by default if none are otherwise specified.
-    static func defaultConnectivityURLs(shouldUseHTTPS: Bool) -> [URL] {
+    static func defaultConnectivityURLRequests(shouldUseHTTPS: Bool) -> [URLRequest] {
         var result: [URL] = []
         let connectivityURLs: [String] = shouldUseHTTPS
         ? ["https://www.apple.com/library/test/success.html",
@@ -449,7 +447,9 @@ private extension Connectivity {
                 result.append(connectivityURL)
             }
         }
-        return result
+        return result.map {
+            URLRequest(url: $0)
+        }
     }
     
     /// Ensures that the instance var `urlSessionConfiguration` is used to create a session.