+++ Connectivity/Classes/Response Validation/ConnectivityResponseValidator.swift
@@ -15,8 +15,8 @@ typealias ResponseValidator = ConnectivityResponseValidator // For internal use.
     /// Determines whether or not the response is valid
     /// and expected for a given `URL`
     ///
-    /// - Parameter url: The `URL`, from which the response was fetched
-    /// - Parameter response: The `URLResponse` returned by url
-    /// - Parameter data: The data in the response returned by url
-    func isResponseValid(url: URL, response: URLResponse?, data: Data?) -> Bool
+    /// - Parameter url: The `URLRequest`, from which the response was fetched.
+    /// - Parameter response: The `URLResponse` returned by url.
+    /// - Parameter data: The data in the response returned by url.
+    func isResponseValid(urlRequest: URLRequest, response: URLResponse?, data: Data?) -> Bool
 }